+++ src/core/tire/tire.h
@@ -27,7 +27,7 @@ class Tire
         enum { end = state_start };
     };
 
-    static_assert(input_names::end == state_names::end);
+    static_assert(static_cast<size_t>(input_names::end) == static_cast<size_t>(state_names::end));
 
     //! Indices of the control variables of this class: none
     struct control_names
